Crate scylla_cql

Re-exports

pub use crate::frame::response::cql_to_rust;

pub use crate::frame::response::cql_to_rust::FromRow;

Deprecated

pub use crate::frame::types::Consistency;

Modules

frame

macros

types

Macros

impl_from_cql_value_from_methodDeprecated

This macro implements FromCqlVal given a type and method of CqlValue that returns this type.

impl_serialize_row_via_value_listDeprecated

Implements the SerializeRow trait for a type, provided that the type already implements the legacy ValueList trait.

impl_serialize_value_via_valueDeprecated

Implements the SerializeValue trait for a type, provided that the type already implements the legacy Value trait.