scylla_cql::deserialize::row

Trait DeserializeRow

pub trait DeserializeRow<'frame, 'metadata>
where
    Self: Sized,
{
    // Required methods
    fn type_check(specs: &[ColumnSpec<'_>]) -> Result<(), TypeCheckError>;
    fn deserialize(
        row: ColumnIterator<'frame, 'metadata>,
    ) -> Result<Self, DeserializationError>;
}

A type that can be deserialized from a row that was returned from a query.

For tips on how to write a custom implementation of this trait, see the documentation of the parent module.

The crate also provides a derive macro which allows to automatically implement the trait for a custom type. For more details on what the macro is capable of, see its documentation.

Required Methods

fn type_check(specs: &[ColumnSpec<'_>]) -> Result<(), TypeCheckError>

Checks that the schema of the result matches what this type expects.

This function can check whether column types and names match the expectations.

fn deserialize( row: ColumnIterator<'frame, 'metadata>, ) -> Result<Self, DeserializationError>

Deserializes a row from given column iterator.

This function can assume that the driver called type_check to verify the row’s type. Note that deserialize is not an unsafe function, so it should not use the assumption about type_check being called as an excuse to run unsafe code.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.