scylla_cql::deserialize

Module value

Provides types for dealing with CQL value deserialization.

Structs

BuiltinDeserializationError

Deserialization of one of the built-in types failed.

BuiltinTypeCheckError

Type checking of one of the built-in types failed.

BytesSequenceIterator

Iterates over a sequence of [bytes] items from a frame subslice.

FixedLengthBytesSequenceIterator

Iterates over a sequence of [bytes] items from a frame subslice, expecting a particular number of items.

ListlikeIterator

An iterator over either a CQL set or list.

MapIterator

An iterator over a CQL map.

UdtIterator

An iterator over fields of a User Defined Type.

VectorIterator

A deserialization iterator over a CQL vector.

Enums

BuiltinDeserializationErrorKind

Describes why deserialization of some of the built-in types failed.

BuiltinTypeCheckErrorKind

Describes why type checking some of the built-in types failed.

MapDeserializationErrorKind

Describes why deserialization of a map type failed.

MapTypeCheckErrorKind

Describes why type checking of a map type failed.

MaybeEmpty

A value that may be empty or not.

SetOrListDeserializationErrorKind

Describes why deserialization of a set or list type failed.

SetOrListTypeCheckErrorKind

Describes why type checking of a set or list type failed.

TupleDeserializationErrorKind

Describes why deserialization of a tuple failed.

TupleTypeCheckErrorKind

Describes why type checking of a tuple failed.

UdtDeserializationErrorKind

Describes why deserialization of a user defined type failed.

UdtTypeCheckErrorKind

Describes why type checking of a user defined type failed.

VectorDeserializationErrorKind

Describes why deserialization of a vector type failed.

VectorTypeCheckErrorKind

Describes why type checking a vector type failed.

Traits

DeserializeValue

A type that can be deserialized from a column value inside a row that was returned from a query.

Emptiable

Values that may be empty or not.