scylla_cql

Module frame

Abstractions of the CQL wire protocol:

request and response frames’ representation and ser/de;

frame header and body;

serialization and deserialization of low-level CQL protocol types;

protocol features negotiation;

compression, tracing, custom payload support;

consistency levels;

errors that can occur during the above operations.

Modules

flag

Frame flags

frame_errors

Low-level errors that can occur during CQL frame parsing and serialization.

protocol_features

Implementation and negotiation of extensions to the CQL protocol.

request

CQL requests sent by the client.

response

CQL responses sent by the server.

server_event_type

This module defines the EventType enum, which represents different types of CQL events.

types

CQL binary protocol in-wire types.

Structs

CompressionFromStrError

Unknown compression.

FrameParams

Parts of the frame header which are not determined by the request/response type.

ResponseBodyWithExtensions

Represents the already parsed response body extensions, including trace ID, warnings, and custom payload, and the remaining body raw data.

SerializedRequest

A serialized CQL request frame, nearly ready to be sent over the wire.

TryFromPrimitiveError

An error type for parsing an enum value from a primitive.

Enums

Authenticator

All of the Authenticators supported by ScyllaDB

Compression

The wire protocol compression algorithm.

Functions

compress_append

Compresses the request body using the specified compression algorithm, appending the compressed data to the provided output buffer.

decompress

Deompresses the response body using the specified compression algorithm and returns the decompressed data as an owned buffer.

parse_response_body_extensions

Decompresses the response body if compression is enabled, and parses any extensions like trace ID, warnings, and custom payload.

read_response_frame

Reads a response frame from the provided reader (usually, a socket). Then parses and validates the frame header and extracts the body.