scylla_cql::frame

Module request

CQL requests sent by the client.

Re-exports

pub use auth_response::AuthResponse;

pub use batch::Batch;

pub use execute::Execute;

pub use options::Options;

pub use prepare::Prepare;

pub use query::Query;

pub use startup::Startup;

Modules

auth_response

CQL protocol-level representation of a AUTH_RESPONSE request.

batch

CQL protocol-level representation of a BATCH request.

execute

CQL protocol-level representation of a EXECUTE request.

options

CQL protocol-level representation of a OPTIONS request.

prepare

CQL protocol-level representation of a PREPARE request.

query

CQL protocol-level representation of a QUERY request.

register

CQL protocol-level representation of a REGISTER request.

startup

CQL protocol-level representation of a STARTUP request.

Enums

CqlRequestKind

Possible requests sent by the client.

Request

A CQL request that can be sent to the server.

RequestOpcode

Opcode of a request, used to identify the request type in a CQL frame.

Traits

DeserializableRequest

Requests that can be deserialized from a CQL frame.

SerializableRequest

Requests that can be serialized into a CQL frame.