scylla_cql::frame::request

Enum Request

#[non_exhaustive]
pub enum Request<'r> {
    Query(Query<'r>),
    Execute(Execute<'r>),
    Batch(Batch<'r, BatchStatement<'r>, Vec<SerializedValues>>),
}

A CQL request that can be sent to the server.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Query(Query<'r>)

QUERY request, used to execute a single unprepared statement.

Execute(Execute<'r>)

EXECUTE request, used to execute a single prepared statement.

Batch(Batch<'r, BatchStatement<'r>, Vec<SerializedValues>>)

BATCH request, used to execute a batch of (prepared, unprepared, or mix of both) statements.

Implementations

impl Request<'_>

pub fn deserialize( buf: &mut &[u8], opcode: RequestOpcode, ) -> Result<Self, RequestDeserializationError>

Deserializes the request from the provided buffer.

pub fn get_consistency(&self) -> Option<Consistency>

Retrieves consistency from request frame, if present.

pub fn get_serial_consistency(&self) -> Option<Option<SerialConsistency>>

Retrieves serial consistency from request frame.