scylla_cql::frame::request::startup

Struct Startup

pub struct Startup<'a> {
    pub options: HashMap<Cow<'a, str>, Cow<'a, str>>,
}

The CQL protocol-level representation of an STARTUP request, used to finalise connection negotiation phase and establish the CQL connection.

Fields
options: HashMap<Cow<'a, str>, Cow<'a, str>>

The protocol options that were suggested by the server and accepted by the client.