scylla_cql::frame

Module response

CQL responses sent by the server.

Re-exports

pub use error::Error;

pub use supported::Supported;

Modules

authenticate

CQL protocol-level representation of an AUTHENTICATE response.

custom_type_parser

Implementation of a parser for custom types in the CQL protocol.

error

CQL protocol-level representation of an ERROR response.

event

CQL protocol-level representation of an EVENT response.

result

CQL protocol-level representation of a RESULT response.

supported

CQL protocol-level representation of a SUPPORTED response.

Enums

CqlResponseKind

Possible CQL responses received from the server

NonErrorResponse

A CQL response that has been received from the server, excluding error responses. This is used to handle responses that are not errors, allowing for easier processing of valid responses without need to handle error case any later.

Response

A CQL response that has been received from the server.

ResponseOpcode

Opcode of a response, used to identify the response type in a CQL frame.