scylla_cql::frame::response

Module result

Structs

ColumnSpec

DeserializedMetadataAndRawRows

RESULT:Rows response, in partially serialized form.

PartitionKeyIndex

Prepared

PreparedMetadata

RawMetadataAndRawRows

RESULT:Rows response, in partially serialized form.

ResultMetadata

SchemaChange

SelfBorrowedMetadataContainer

A container that can be considered an Arc<ResultMetadata> with an additional capability of containing metadata in a borrowed form.

SetKeyspace

TableSpec

UserDefinedType

Definition of a user-defined type

Enums

CollectionType

Collection variants of ColumnType. A collection is a composite type that has dynamic size, so it is possible to add and remove values to/from it.

ColumnType

A type of:

NativeType

A ColumnType variants that are “simple” (non-recursive).

Result

ResultMetadataHolder

Versatile container for ResultMetadata. Allows 2 types of ownership of ResultMetadata:

Functions

deserialize