scylla_cql::frame

Module types

CQL binary protocol in-wire types.

Structs

NonSerialConsistencyError

Error returned when a serial consistency what expected, yet got another kind of consistency.

Enums

Consistency

A setting that defines a successful write or read by the number of cluster replicas that acknowledge the write or respond to the read request, respectively. See ScyllaDB docs for more detailed description and guidelines.

RawValue

SerialConsistency

Consistency for Lightweight Transactions (LWTs).

Functions

read_bytes

read_bytes_map

read_bytes_opt

read_consistency

read_inet

read_int

read_int_length

read_long

read_long_string

read_short

read_short_bytes

read_string

read_string_list

read_string_map

read_string_multimap

read_uuid

read_value

write_bytes

write_bytes_map

write_bytes_opt

write_consistency

write_inet

write_int

write_long

write_long_string

write_serial_consistency

write_short

write_short_bytes

write_string

write_string_list

write_string_map

write_string_multimap

write_uuid