Module scylla_cql::frame::value

Structs

Counter

Represents an counter value

CqlDate

Native CQL date representation that allows for a bigger range of dates (-262145-1-1 to 262143-12-31).

CqlDecimal

Native CQL decimal representation.

CqlDuration

Represents a CQL Duration value

CqlTime

Native CQL time representation.

CqlTimestamp

Native CQL timestamp representation that allows full supported timestamp range.

CqlTimeuuid

Represents timeuuid (uuid V1) value

CqlVarint

Native CQL varint representation.

LegacyBatchValuesFirstSerialized

Allows reusing already-serialized first value

LegacyBatchValuesFromIter

Implements BatchValues from an Iterator over references to things that implement ValueList

LegacyBatchValuesIteratorFromIterator

Implements BatchValuesIterator from an Iterator over references to things that implement ValueList

LegacySerializedValues

Keeps a buffer with serialized Values Allows adding new Values and iterating over serialized ones

LegacySerializedValuesIterator

TupleValuesIter

Unset

Represents an unset value

ValueOverflow

ValueTooBig

Enums

MaybeUnset

Enum providing a way to represent a value that might be unset

SerializeValuesError

Traits

LegacyBatchValues

Represents List of ValueList for Batch statement

LegacyBatchValuesIterator

An iterator-like for ValueList

Value

Every value being sent in a query must implement this trait serialize() should write the Value as bytes to the provided buffer

ValueList

Represents list of values to be sent in a query gets serialized and but into request

Type Aliases

SerializedResult