Struct scylla_cql::frame::value::CqlDecimal

pub struct CqlDecimal { /* private fields */ }

Native CQL decimal representation.

Represented as a pair:

a CqlVarint value

32-bit integer which determines the position of the decimal point

The type is not very useful in most use cases. However, users can make use of more complex types such as bigdecimal::BigDecimal (v0.4). The library support (e.g. conversion from CqlValue) for the type is enabled via bigdecimal-04 crate feature.

DB data format

Notice that constructors donโt perform any normalization on the provided data. For more details, see CqlVarint documentation.

Implementations

impl CqlDecimal

Constructors

pub fn from_signed_be_bytes_and_exponent(bytes: Vec<u8>, scale: i32) -> Self

Creates a CqlDecimal from an array of bytes representing CqlVarint and a 32-bit scale.

See: disclaimer about non-normalized values.

pub fn from_signed_be_bytes_slice_and_exponent(bytes: &[u8], scale: i32) -> Self

Creates a CqlDecimal from a slice of bytes representing CqlVarint and a 32-bit scale.

See: disclaimer about non-normalized values.

impl CqlDecimal

Conversion to raw bytes

pub fn as_signed_be_bytes_slice_and_exponent(&self) -> (&[u8], i32)

Returns a slice of bytes in twoโs complement binary big-endian representation and a scale.

pub fn into_signed_be_bytes_and_exponent(self) -> (Vec<u8>, i32)

Converts CqlDecimal to an array of bytes in twoโs complement binary big-endian representation and a scale.