scylla_cql
Module macros

Macros

impl_from_cql_value_from_methodDeprecated

This macro implements FromCqlVal given a type and method of CqlValue that returns this type.

impl_serialize_row_via_value_listDeprecated

Implements the SerializeRow trait for a type, provided that the type already implements the legacy ValueList trait.

impl_serialize_value_via_valueDeprecated

Implements the SerializeValue trait for a type, provided that the type already implements the legacy Value trait.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

Traits

BufMut

A trait for values that provide sequential write access to bytes.

Derive Macros

DeserializeRow

DeserializeValue

FromRow

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.

FromUserType

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.

IntoUserType

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.

SerializeRow

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.

SerializeValue

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.

ValueList

Documentation for this macro can only be found in scylla crate - not in scylla-macros nor in scylla-cql. This is because of rustdocs limitations that are hard to explain here.