scylla_cql::serialize::raw_batch

Trait RawBatchValues

pub trait RawBatchValues {
    type RawBatchValuesIter<'r>: RawBatchValuesIterator<'r>
       where Self: 'r;

    // Required method
    fn batch_values_iter(&self) -> Self::RawBatchValuesIter<'_>;
}

Represents a list of sets of values for a batch statement.

Unlike BatchValues), it doesn’t require type information from the statements of the batch in order to be serialized.

This is a lower level trait than BatchValues) and is only used for interaction between the code in scylla and scylla-cql crates. If you are a regular user of the driver, you shouldn’t care about this trait at all.

Required Associated Types

type RawBatchValuesIter<'r>: RawBatchValuesIterator<'r> where Self: 'r

An Iterator-like object over the values from the parent BatchValues object.

Required Methods

fn batch_values_iter(&self) -> Self::RawBatchValuesIter<'_>

Returns an iterator over the data contained in this object.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.