Module scylla_cql::types::serialize

Types and traits related to serialization of values to the CQL format.

Re-exports

pub use writers::CellValueBuilder;

pub use writers::CellWriter;

pub use writers::RowWriter;

Modules

batch

Contains the BatchValues and BatchValuesIterator trait and their implementations.

raw_batch

Contains the RawBatchValues and RawBatchValuesIterator trait and their implementations.

row

Contains the SerializeRow trait and its implementations.

value

Contains the SerializeCql trait and its implementations.

writers

Contains types and traits used for safe serialization of values for a CQL statement.

Structs

SerializationError

An error indicating that a failure happened during serialization.