Module scylla_cql::types::serialize::raw_batch

Contains the RawBatchValues and RawBatchValuesIterator trait and their implementations.

Structs

RawBatchValuesAdapter

Takes BatchValues and an iterator over contexts, and turns them into a RawBatchValues.

RawBatchValuesIteratorAdapter

Takes BatchValuesIterator and an iterator over contexts, and turns them into a RawBatchValuesIterator.

Traits

RawBatchValues

Represents a list of sets of values for a batch statement.

RawBatchValuesIterator

An Iterator-like object over the values from the parent RawBatchValues object.