scylla_cql::types::serialize
Module value

Contains the SerializeValue trait and its implementations.

Structs

BuiltinSerializationError

Serialization of one of the built-in types failed.

BuiltinTypeCheckError

Type checking of one of the built-in types failed.

ValueAdapterDeprecated

Implements SerializeValue if the type wrapped over implements Value.

Enums

BuiltinSerializationErrorKind

Describes why serialization of some of the built-in types has failed.

BuiltinTypeCheckErrorKind

Describes why type checking some of the built-in types has failed.

MapSerializationErrorKind

Describes why serialization of a map type failed.

MapTypeCheckErrorKind

Describes why type checking of a map type failed.

SetOrListSerializationErrorKind

Describes why serialization of a set or list type failed.

SetOrListTypeCheckErrorKind

Describes why type checking of a set or list type failed.

TupleSerializationErrorKind

Describes why serialize of a tuple failed.

TupleTypeCheckErrorKind

Describes why type checking of a tuple failed.

UdtSerializationErrorKind

Describes why serialization of a user defined type failed.

UdtTypeCheckErrorKind

Describes why type checking of a user defined type failed.

ValueToSerializeValueAdapterErrorDeprecated

Describes a failure to translate the output of the Value legacy trait into an output of the SerializeValue trait.

Traits

SerializeValue

A type that can be serialized and sent along with a CQL statement.

Functions

serialize_legacy_valueDeprecated

Serializes a value implementing Value by using the CellWriter interface.