Enum scylla_cql::types::serialize::value::UdtTypeCheckErrorKind

#[non_exhaustive]
pub enum UdtTypeCheckErrorKind {
    NotUdt,
    NameMismatch {
        keyspace: String,
        type_name: String,
    },
    ValueMissingForUdtField {
        field_name: String,
    },
    NoSuchFieldInUdt {
        field_name: String,
    },
    FieldNameMismatch {
        rust_field_name: String,
        db_field_name: String,
    },
}

Describes why type checking of a user defined type failed.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NotUdt

The CQL type is not a user defined type.

NameMismatch

The name of the UDT being serialized to does not match.

Fields

keyspace: String

Keyspace in which the UDT was defined.

type_name: String

Name of the UDT.

ValueMissingForUdtField

The Rust data does not have a field that is required in the CQL UDT type.

Fields

field_name: String

Name of field that the CQL UDT requires but is missing in the Rust struct.

NoSuchFieldInUdt

The Rust data contains a field that is not present in the UDT.

Fields

field_name: String

Name of the Rust struct field that is missing in the UDT.

FieldNameMismatch

A different field name was expected at given position.

Fields

rust_field_name: String

The name of the Rust field.

db_field_name: String

The name of the CQL UDT field.