Crate scylla_cql

Defines CQL protocol-level types and traits for interacting with ScyllaDB/Cassandra.

Mainly intended to be used by the ScyllaDB driver, but can also be useful for other applications that need to interact with CQL.

Re-exports

pub use crate::frame::types::Consistency;

Modules

deserialize

Framework for deserialization of data returned by database queries.

frame

Abstractions of the CQL wire protocol:

serialize

Types and traits related to serialization of values to the CQL format.

utils

Various utilities.

value

Defines CQL values of various types and their representations, as well as conversion between them and other types.

Derive Macros

DeserializeRow

Derive macro for the DeserializeRow trait that generates an implementation which deserializes a row with a similar layout to the Rust struct.

DeserializeValue

Derive macro for the DeserializeValue trait that generates an implementation which deserializes a User Defined Type with the same layout as the Rust struct.

SerializeRow

Derive macro for the SerializeRow trait which serializes given Rust structure into bind markers for a CQL statement.

SerializeValue

Derive macro for the SerializeValue trait which serializes given Rust structure as a User Defined Type (UDT).