Crate scylla_cql

Re-exports

pub use crate::frame::response::cql_to_rust;

pub use crate::frame::response::cql_to_rust::FromRow;

pub use crate::frame::types::Consistency;

Modules

errors

This module contains various errors which can be returned by scylla::Session

frame

macros

types

Macros

impl_from_cql_value_from_method

This macro implements FromCqlVal given a type and method of CqlValue that returns this type.

impl_serialize_cql_via_value

Implements the SerializeCql trait for a type, provided that the type already implements the legacy Value trait.

impl_serialize_row_via_value_list

Implements the SerializeRow trait for a type, provided that the type already implements the legacy ValueList trait.