Module scylla_cql::frame::frame_errors

Re-exports

pub use super::request::auth_response::AuthResponseSerializationError;

pub use super::request::batch::BatchSerializationError;

pub use super::request::batch::BatchStatementSerializationError;

pub use super::request::execute::ExecuteSerializationError;

pub use super::request::prepare::PrepareSerializationError;

pub use super::request::query::QueryParametersSerializationError;

pub use super::request::query::QuerySerializationError;

pub use super::request::register::RegisterSerializationError;

pub use super::request::startup::StartupSerializationError;

Structs

ColumnSpecParseError

An error type returned when deserialization of table column specifications fails.

Enums

ClusterChangeEventParseError

An error type returned when deserialization of [Status/Topology]ChangeEvent fails.

ColumnSpecParseErrorKind

The type of error that appeared during deserialization of a column specification.

CqlAuthChallengeParseError

An error type returned when deserialization of AUTH_CHALLENGE response fails.

CqlAuthSuccessParseError

An error type returned when deserialization of AUTH_SUCCESS response fails.

CqlAuthenticateParseError

An error type returned when deserialization of AUTHENTICATE response fails.

CqlErrorParseError

An error type returned when deserialization of ERROR response fails.

CqlEventParseError

An error type returned when deserialization of EVENT response fails.

CqlRequestSerializationError

An error that occurred during CQL request serialization.

CqlResponseParseError

An error type returned when deserialization of CQL server response fails.

CqlResultParseError

An error type returned when deserialization of RESULT response fails.

CqlSupportedParseError

An error type returned when deserialization of SUPPORTED response fails.

CqlTypeParseError

An error type returned when deserialization of CQL type name fails.

FrameBodyExtensionsParseError

An error returned by parse_response_body_extensions.

FrameHeaderParseError

An error that occurred during frame header deserialization.

LowLevelDeserializationError

A low level deserialization error.

PreparedMetadataParseError

An error type returned when deserialization of statement’s prepared metadata failed.

PreparedParseError

An error type returned when deserialization of RESULT::Prepared` response fails.

RawRowsAndPagingStateResponseParseError

An error that occurred during initial deserialization of RESULT:Rows response. Since the deserialization of rows is lazy, we initially only need to deserialize:

ResultMetadataAndRowsCountParseError

An error returned when lazy deserialization of result metadata and rows count fails.

ResultMetadataParseError

An error type returned when deserialization of result metadata failed.

SchemaChangeEventParseError

An error type returned when deserialization of SchemaChangeEvent fails.

SetKeyspaceParseError

TableSpecParseError

An error type returned when deserialization of table specification fails.