scylla_cql::frame::response

Module result

CQL protocol-level representation of a RESULT response.

Structs

ColumnSpec

Specification of a column of a table.

DeserializedMetadataAndRawRows

RESULT:Rows response, in partially serialized form.

PartitionKeyIndex

Represents the relationship between partition key columns and bind markers. This allows implementations with token-aware routing to correctly construct the partition key without needing to inspect table metadata.

Prepared

Represents a CQL RESULT::Prepared response, which is sent in response to a PREPARE request.

PreparedMetadata

Metadata of a prepared statement about its bound values.

RawMetadataAndRawRows

RESULT:Rows response, in partially serialized form.

ResultMetadata

Metadata of a result set.

SchemaChange

Represents a CQL RESULT::SchemaChange response, which indicates that a schema changed as an effect of the executed request.

SelfBorrowedMetadataContainer

A container that can be considered an Arc<ResultMetadata> with an additional capability of containing metadata in a borrowed form.

SetKeyspace

Represents a CQL RESULT::SetKeyspace response.

TableSpec

Specification of a table in a keyspace.

UserDefinedType

Definition of a user-defined type (UDT). UDT is composed of fields, each with a name and an optional value of its own type.

Enums

CollectionType

Collection variants of ColumnType. A collection is a composite type that has dynamic size, so it is possible to add and remove values to/from it.

ColumnType

A type of:

NativeType

A ColumnType variants that are “simple” (non-recursive).

Result

Represents the result of a CQL RESULT response.

ResultMetadataHolder

Versatile container for ResultMetadata. Allows 2 types of ownership of ResultMetadata:

Functions

deserialize

Deserializes a CQL RESULT response from the provided buffer.