scylla_cql::frame::value
Trait Value

pub trait Value {
    // Required method
    fn serialize(&self, buf: &mut Vec<u8>) -> Result<(), ValueTooBig>;
}

👎Deprecated since 0.15.1: Legacy serialization API is not type-safe and is going to be removed soon

Every value being sent in a query must implement this trait serialize() should write the Value as bytes to the provided buffer

Required Methods

fn serialize(&self, buf: &mut Vec<u8>) -> Result<(), ValueTooBig>

👎Deprecated since 0.15.1: Legacy serialization API is not type-safe and is going to be removed soon